+++ src/integer64.c
@@ -20,6 +20,7 @@
 #include <math.h> // floor
 #include <stdint.h>
 #include <stdlib.h> // strtoll
+#include <stdbool.h> // for boolean
 
 #include <R.h>
 #include <Rdefines.h>
@@ -680,12 +681,16 @@ SEXP min_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MAX_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
-            if (e1[i]!=NA_INTEGER64 && e1[i]<ret[0]){
+            if (e1[i]!=NA_INTEGER64){
+              onlyNas = false;
+              if (e1[i]<ret[0]){
                 ret[0] = e1[i];
-            }
+              }
+            } 
         }
     }else{
         for(i=0; i<n; i++){
@@ -693,11 +698,15 @@ SEXP min_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;       
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+  }
   return ret_;
 }
 
@@ -705,12 +714,16 @@ SEXP max_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MIN_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
-            if (e1[i]!=NA_INTEGER64 && e1[i]>ret[0]){
+            if (e1[i]!=NA_INTEGER64){
+              onlyNas = false;
+              if (e1[i]>ret[0]){
                 ret[0] = e1[i];
-            }
+              }
+            } 
         }
     }else{
         for(i=0; i<n; i++){
@@ -718,11 +731,15 @@ SEXP max_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;       
                 if (e1[i]>ret[0])
                     ret[0] = e1[i];
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+  }
   return ret_;
 }
 
@@ -730,11 +747,13 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
   long long i, n = LENGTH(e1_);
   long long * e1 = (long long *) REAL(e1_);
   long long * ret = (long long *) REAL(ret_);
+  bool onlyNas = true;
   ret[0] = MAX_INTEGER64;
   ret[1] = MIN_INTEGER64;
     if (asLogical(na_rm_)){
         for(i=0; i<n; i++){
             if (e1[i]!=NA_INTEGER64){
+                onlyNas = false;
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
                 if (e1[i]>ret[1])
@@ -747,6 +766,7 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
                 ret[0] = ret[1] = NA_INTEGER64;
                 return ret_;
             }else{
+                onlyNas = false;
                 if (e1[i]<ret[0])
                     ret[0] = e1[i];
                 if (e1[i]>ret[1])
@@ -754,6 +774,10 @@ SEXP range_integer64(SEXP e1_, SEXP na_rm_, SEXP ret_){
             }
         }
     }
+  if (onlyNas){
+    ret[0] = NA_INTEGER64;
+    ret[1] = NA_INTEGER64;
+  }
   return ret_;
 }